unncessary unboxing
Explicit unwrapping of wrapped primitive values. Unboxing is unnecessary
under Java 5 and newer, and can be safely removed.

+++ tools/cooja/java/org/contikios/cooja/Cooja.java
@@ -3760,7 +3760,7 @@ public class Cooja extends Observable {
           if (plugin.getClientProperty("zorder") == null) {
           	continue;
           }
-          int zOrder = ((Integer) plugin.getClientProperty("zorder")).intValue();
+          int zOrder = (Integer) plugin.getClientProperty("zorder");
           if (zOrder != z) {
           	continue;
           }
